Refactor and simplify test code
- Use os.paths directly.
- Move recording tests from test_flask to test.web_framework.
